# Optimise dumping to reduce unnecessary overhead