@dsimidzija 2020-08-19
I didn't even know about #805, good to know! My ideas were something similar: figure out a way to cache values for certain types at least. For example, if I know that a specific field is a string property and it is 100% there, there should be no reason to have complicated lookups or enforcing types. I didn't have time to delve deeper into the various Field subclasses, but `line_profiler` implies a lot of repetition there.

As I'm typing this, I wonder if it would be possible to do something like [lima is doing](https://github.com/b6d/lima/blob/develop/lima/schema.py#L76), and compile a single function which dumps the entire object, where each Field subclass could in theory produce the "most optimized" serializer for itself.