@dsimidzija 2021-03-16
I don't know if this is of any interest, but [with some minimal changes](https://github.com/dsimidzija/marshmallow/commit/fe8d898f7304c60b658ef609dd326e07cfcaaef7), marshmallow can be cythonized with [setuptools-cythonize](https://pypi.org/project/setuptools-cythonize/) and together with this MR, the performance is around ~45% better!