@lafrech 2021-03-16
I'm not familiar with cynthonize. Does cynthonization itself - without this PR - bring a significative improvement?

Is there anything we could do to make it available more easily? Like, could/should we distribute binary packages? Would it be useful to many users or is it a niche?