@sloria 2025-01-14
closing this for now as it's stale. let's continue discussion on #805 . nice work on https://github.com/dsimidzija/python-althaia btw!